Origins updates, some models, bugfixes
I am so sorry to anyone who wants a descriptive comment. There's so much shit in every update and I am sublimely lazy

+++ src/main/java/fun/mousewich/block/BlazePowderBlock.java
@@ -54,7 +54,7 @@ public class BlazePowderBlock extends ModFallingBlock {
 		BlockPos blockPos = pos.up();
 		if (world.getFluidState(pos).isIn(FluidTags.WATER)) {
 			world.playSound(null, pos, SoundEvents.BLOCK_FIRE_EXTINGUISH, SoundCategory.BLOCKS, 0.5f, 2.6f + (world.random.nextFloat() - world.random.nextFloat()) * 0.8f);
-			world.spawnParticles(ParticleTypes.LARGE_SMOKE, (double)blockPos.getX() + 0.5, (double)blockPos.getY() + 0.25, (double)blockPos.getZ() + 0.5, 8, 0.5, 0.25, 0.5, 0.0);
+			world.spawnParticles(ParticleTypes.LARGE_SMOKE, blockPos.getX() + 0.5, blockPos.getY() + 0.25, blockPos.getZ() + 0.5, 8, 0.5, 0.25, 0.5, 0.0);
 		}
 	}
 }
